Fix #3388
Overflow: hidden was cutting of info in the play view. Without it in the new home page, though,
infinite banners would expand the size of the page. Tried figuring out a way to not make the new
page require hiding overflow but found no obvious solution. Pages with infinite banners will
just need to be set to overflow: hidden individually.