Add third-party API
This is the basis for third parties to incorporate CodeCombat as part of their
curriculums and programs. Third parties can:

* Create users
* Give OAuth2 identities
* Grant premium access
* Fetch progress info about the users

Since third parties will both be interacting with CodeCombat as an API Client
and logging in their users as OAuth2 Providers, these roles are in two separate
collections: `Client` and `OAuthProvider`. The `OAuthProvider` collection
can also eventually be used as the basis for existing identity integrations we
have: Facebook, G+ and Clever.

Prepaids have been slightly extended to serve as a record of 'purchases' by
third parties. They're currently used for buying premium access for other users
(type `'terminal_subscription'`) and for licenses bought by schools for course
access (type `'course'`). The former is for a length of time (months), the latter
is for a specific time frame (from date to date). With this new version, the
`'terminal_subscription'` type can now be for a specific time frame, and that's
used also to keep track of how much has been 'spent' by these third parties.
I also added `clientCreator` to be clear when a prepaid is made by an API Client
as opposed to a user.

Other changes:
* Refactored POST /db/subscription/-/subscribe_prepaid
* Removed toObject TODOs because I tried having APIClient require req, and it was a nuisance in practice
* Ensure User indexes finish building for tests
* Promisify and correct callback behavior for a couple server functions
* Add middleware for special logging all network errors for a given endpoint
* Add 'subscription' virtual property to server User