Allow saving user settings when starting off invalid
Sometimes properties are added and get ahead of schema changes.
In these cases, users cannot edit their settings because schema
validation fails; both server and client throw an error. This commit
tweaks both so that if the user is in an invalid state, an error only
occurs if there are *new* schema errors.

Also refactored, cleaned up some user tests.