Prevent timed out requests from throwing errors when they try to respond
Even if the timedout middleware finishes the response, the logic
that's taking a while can still finish and try to send a response.
If this happens, the resulting error does NOT get handled by
the error middleware, and just crashes the worker instead.
So make the replace all response-ending functions on the
response with a chainable noop, so res.status(200).send('OK') works.