Fix BuyGemsModal translations
Two years ago, I forgot to add an i18n property to the product schema, which
the BuyGemsModal depended on to translate labels. This hidden property
was a string that pointed to the static translation.

https://github.com/codecombat/codecombat/commit/7c516c4d9fe334af8db3fc0d8fe8840dca543519

More recently, we added the ability to translate products for PayPal
payments. This created an i18n property, but an object, for database-stored
translations.

https://github.com/codecombat/codecombat/commit/b02069e33757e4d55ccfb7b014b167f2bf13ced1

This broke the BuyGemsModal. It didn't really make sense to store
translation strings specifically for the BuyGemsModal anyway, so I just
have that view handle relating gem product names to descriptions.