Fix jade crashing webpack
There was a fix in the latest version of pug-loader, but upgrading
isn't simple.

Tried moving everything to the latest pug2 rc, various templates
broke (including world map) because they rely on interpolation in
attributes and that version of pug2 doesn't support that feature.

Tried having both jade and pug for a gradual migration but the static
files depend on the older version of pug.

Ended up forking our own version of pug-loader v0.8 and putting in
the fix manually for it.