Fix race condition in CampaignView
Sometimes, the poll request will return 404, and when that happens
the view stops listening to the supermodel and it never really 'loads',
and looks as if the user did not have any sessions. Changed the
'stop listening' behavior to only happen for blocking requests, ie
ones with value greater than 0.